requirements.txt
appnope==0.1.0
asn1crypto==0.23.0
awscli==1.14.6
backports.shutil-get-terminal-size==1.0.0
boto==2.48.0
boto3==1.4.8
botocore==1.8.10
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.7
confusable-homoglyphs==2.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
decorator==4.1.2
Django==1.11.7
django-filter==1.1.0
django-registration==2.3
django-rest-swagger==2.1.2
django-transfer==0.4
djangorestframework==3.7.3
docutils==0.14
-e git+https://github.com/manosim/django-rest-framework-docs.git@94571d05617d24816ba8fdc5e836d17645a88319#egg=drfdocs
enum34==1.1.6
futures==3.2.0
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.18
ipython==5.5.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
Markdown==2.6.9
MarkupSafe==1.0
oauthlib==2.0.6
openapi-codec==1.3.2
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.12
scandir==1.6
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
